+++ docs/gatsby-config.js
@@ -2,11 +2,13 @@ const Path = require('path')
 module.exports = {
   pathPrefix: `/wizard-ui`,
   siteMetadata: {
+    siteUrl: 'https://wizard-ui.netlify.com/',
     title: `Wizard UI Documentation`,
     description: `Wizard UI Documentation.`,
     author: `xsky contributors`,
   },
   plugins: [
+    `gatsby-plugin-sitemap`,
     {
       resolve: `gatsby-source-filesystem`,
       options: {
@@ -33,6 +35,22 @@ module.exports = {
     `gatsby-transformer-sharp`,
     `gatsby-plugin-sharp`,
     {
+      resolve: `@gatsby-contrib/gatsby-plugin-elasticlunr-search`,
+      options: {
+        // Fields to index
+        fields: ['title', 'author', 'path'],
+        // How to resolve each field`s value for a supported node type
+        resolvers: {
+          // For any node of type Mdx, list how to resolve the fields` values
+          Mdx: {
+            title: node => node.frontmatter.title,
+            author: node => node.frontmatter.author,
+            path: node => node.fields.slug,
+          },
+        },
+      },
+    },
+    {
       resolve: `gatsby-plugin-manifest`,
       options: {
         name: `wizard-ui`,
@@ -48,6 +66,9 @@ module.exports = {
       resolve: `gatsby-plugin-sass`,
       cssLoaderOptions: {
         camelCase: false,
+      },
+      options: {
+        precision: 8,
       },
     },
   ],