+++ src/interface.tsx
@@ -1,6 +1,20 @@
 import * as React from 'react';
 import { SelectCallback, Sizes } from 'react-bootstrap';
 import { Moment } from 'moment';
+import CSS from 'csstype';
+
+export interface Query {
+  offset?: number;
+  limit?: number;
+  q?: string;
+}
+
+export interface FetchResponse <T = any> {
+  response?: {
+    [res: string]: T | T[];
+  };
+  error?: string;
+}
 
 export interface BadgeProps {
   count?: number | string;
@@ -61,6 +75,7 @@ export interface TabsProps {
   eventKeyName?: string;
   direction?: 'right';
   onSelect?: SelectCallback;
+  id?: string;
 }
 
 export interface StepsProps {
@@ -78,6 +93,7 @@ export interface ModalProps {
   confirmText?: string;
   okStyle?: string;
   loading?: boolean;
+  hideFooter?: boolean;
 }
 
 interface SwitchInput {
@@ -126,13 +142,20 @@ export interface SubMenuProps {
   children: React.ReactNode;
 }
 
-export interface DropdownButtonMenuItem {
-  key?: string | number;
-  children: DropdownButtonMenuItem[];
-  title: string;
-}
+export type DropdownButtonMenuItem =
+  | {
+      key?: string | number;
+      children?: DropdownButtonMenuItem[];
+      title: string;
+      eventKey?: string;
+      'data-action'?: string;
+    }
+  | string;
 
-export interface DropdownButtonProps {
+export interface DefaultDropdownButtonProps {
+  componentClass: any;
+}
+export interface DropdownButtonProps extends DefaultDropdownButtonProps {
   bsStyle?: string;
   id: string;
   onSelect?: SelectCallback;
@@ -140,8 +163,7 @@ export interface DropdownButtonProps {
   bsSize?: Sizes;
   title?: string | React.ReactNode;
   menu?: DropdownButtonMenuItem[];
-  children: React.ReactNode;
-  componentClass?: any;
+  children?: React.ReactNode;
   modifer?: string;
   disabled?: boolean;
   dropup?: boolean;
@@ -210,6 +232,7 @@ export interface PanelProps {
   children?: React.ReactNode;
   bg?: string;
   text?: string;
+  className?: string;
 }
 
 export interface DropdownDefaultProps {
@@ -220,4 +243,91 @@ export interface DropdownProps extends DropdownDefaultProps {
   className?: string;
   title?: string;
   children?: React.ReactNode;
+  customToggle?: boolean;
+  pullRight?: boolean;
+}
+
+export interface MenuItemOptions {
+  title: string;
+  value: string;
+}
+export interface InputDropdownProps {
+  options?: MenuItemOptions[];
+  defaultValue?: string;
+  value?: string;
+  onChange?: SelectCallback;
+  input?: any;
+  meta?: any;
+}
+export type VirtualItem = {
+  id?: number;
+} | string | number
+export interface VirtualRowArgs<T> {
+  index: number;
+  item: T;
+  prevItem: T | null;
+  nextItem: T | null;
+  style: CSS.Properties;
+}
+export interface VirtualAnchorItem {
+  index: number;
+  offset: number;
+}
+export interface VirtualListState {
+  startIndex: number;
+  endIndex: number;
+}
+export interface VirtualListDefaultProps<T> {
+  height?: number | string;
+  data: T[];
+  runwayItems?: number;
+  runwayItemsOppsite?: number;
+  loader?: React.ReactNode;
+  placeholder?: React.ReactNode | string;
+  noMoreHint?: React.ReactNode | boolean;
+  debug?: boolean;
+}
+export interface VirtualListProps<T> extends VirtualListDefaultProps<T> {
+  query?: Query;
+  onQueryChange?: (query: Query) => Promise<void>;
+  rowHeight?: number | ((item: object) => number);
+  rowRenderer: (item: VirtualRowArgs<T>) => React.ReactNode | Element;
+  isFetching?: boolean;
+  isReloading?: boolean;
+  noMore?: boolean;
+  totalCount?: number;
+  className?: string;
+  isEstimate?: boolean;
+}
+
+export interface VirtualSelectBoxDefaultProps<T> {
+  rowHeight: number;
+  isBtn: boolean;
+  disabled: boolean;
+  placeholder: string;
+  query: Query;
+  defaultItem: T;
+}
+export interface VirtualSelectBoxProps<T> extends VirtualSelectBoxDefaultProps<T> {
+  fetchData: (isReloading: boolean, query: Query, search?: string) => Promise<{
+    query: Query;
+    items: T[];
+    totalCount: number;
+    error?: string;
+  }>;
+  item?: T;
+  className?: string;
+  clear?: boolean;
+  onSelect?: (item: T) => void;
+  formatOption?: (item: T) => T;
+}
+export interface VirtualSelectBoxState<T> {
+  search: string;
+  items: T[];
+  query: Query;
+  isFetching: boolean;
+  totalCount: number;
+  isOpen: boolean;
+  isReloading: boolean;
+  error?: string;
 }